test: Moved schedule unit-test to integration test.
