+++ crates/holochain/src/core/ribosome/host_fn/call_info.rs
@@ -93,7 +93,7 @@ pub fn call_info(
 #[cfg(test)]
 #[cfg(feature = "slow_tests")]
 pub mod test {
-    use crate::core::ribosome::wasm_test::RibosomeTestFixture;
+    use crate::test_utils::RibosomeTestFixture;
     use holochain_wasm_test_utils::TestWasm;
     use holochain_zome_types::prelude::*;
