+++ crates/holochain/src/core/ribosome/host_fn/capability_grants.rs
@@ -16,7 +16,7 @@ pub fn capability_grants(
 #[cfg(test)]
 #[cfg(feature = "slow_tests")]
 pub mod wasm_test {
-    use crate::core::ribosome::wasm_test::RibosomeTestFixture;
+    use crate::test_utils::RibosomeTestFixture;
     use ::fixt::prelude::*;
     use hdk::prelude::*;
     use holochain_types::fixt::CapSecretFixturator;
