+++ crates/holochain/src/core/ribosome/host_fn/delete.rs
@@ -108,7 +108,7 @@ pub(crate) fn get_original_entry_data(
 #[cfg(test)]
 #[cfg(feature = "slow_tests")]
 pub mod wasm_test {
-    use crate::core::ribosome::wasm_test::RibosomeTestFixture;
+    use crate::test_utils::RibosomeTestFixture;
     use hdk::prelude::*;
     use holochain_wasm_test_utils::TestWasm;
