+++ crates/holochain/src/core/ribosome/host_fn/sign_ephemeral.rs
@@ -52,7 +52,7 @@ pub fn sign_ephemeral(
 #[cfg(test)]
 #[cfg(feature = "slow_tests")]
 pub mod wasm_test {
-    use crate::core::ribosome::wasm_test::RibosomeTestFixture;
+    use crate::test_utils::RibosomeTestFixture;
     use hdk::prelude::*;
     use holochain_keystore::AgentPubKeyExt;
     use holochain_wasm_test_utils::TestWasm;
