+++ crates/holochain/src/core/ribosome/host_fn/sys_time.rs
@@ -33,7 +33,7 @@ pub fn sys_time(
 #[cfg(test)]
 #[cfg(feature = "slow_tests")]
 pub mod wasm_test {
-    use crate::core::ribosome::wasm_test::RibosomeTestFixture;
+    use crate::test_utils::RibosomeTestFixture;
     use holochain_wasm_test_utils::TestWasm;
     use holochain_zome_types::prelude::Timestamp;
