+++ crates/holochain/src/core/ribosome/host_fn/x_salsa20_poly1305_encrypt.rs
@@ -52,8 +52,7 @@ pub fn x_salsa20_poly1305_encrypt(
 #[cfg(test)]
 #[cfg(feature = "slow_tests")]
 pub mod wasm_test {
-
-    use crate::core::ribosome::wasm_test::RibosomeTestFixture;
+    use crate::test_utils::RibosomeTestFixture;
     use hdk::prelude::*;
     use holochain_wasm_test_utils::TestWasm;
