+++ crates/holochain/tests/tests/schedule.rs
@@ -1,14 +1,19 @@
 use hdk::prelude::{BoxApi, ExternIO, InlineZomeResult};
 use holochain::sweettest::{SweetCell, SweetConductor, SweetDnaFile};
-use holochain::test_utils::host_fn_caller::HostFnCaller;
+use holochain::test_utils::{host_fn_caller::HostFnCaller, RibosomeTestFixture};
 use holochain_sqlite::error::DatabaseError;
 use holochain_state::mutations::schedule_fn;
+use holochain_state::prelude::*;
 use holochain_state::schedule::fn_is_scheduled;
+use holochain_state::schedule::live_scheduled_fns;
 use holochain_timestamp::Timestamp;
 use holochain_types::inline_zome::InlineZomeSet;
 use holochain_types::signal::Signal;
-use holochain_zome_types::prelude::Schedule::{Ephemeral, Persisted};
-use holochain_zome_types::prelude::{Schedule, ScheduledFn};
+use holochain_wasm_test_utils::TestWasm;
+use holochain_zome_types::prelude::{
+    Schedule::{self, Ephemeral, Persisted},
+    ScheduledFn,
+};
 use holochain_zome_types::signal::AppSignal;
 use tokio::sync::broadcast;
 use tokio::time::error::Elapsed;
@@ -379,4 +384,283 @@ async fn wait_for_signal(
         }
         _ => panic!("Expected AppSignal, got {:?}", msg),
     }
+}
+
+#[tokio::test(flavor = "multi_thread")]
+#[cfg(feature = "test_utils")]
+async fn schedule_test_low_level() -> anyhow::Result<()> {
+    holochain_trace::test_run();
+    let RibosomeTestFixture {
+        alice_pubkey,
+        alice_host_fn_caller,
+        ..
+    } = RibosomeTestFixture::new(TestWasm::Schedule).await;
+
+    alice_host_fn_caller
+        .authored_db
+        .write_async(move |txn| {
+            let now = Timestamp::now();
+            let the_past = (now - std::time::Duration::from_millis(1)).unwrap();
+            let the_future = (now + std::time::Duration::from_millis(1000)).unwrap();
+            let the_distant_future = (now + std::time::Duration::from_millis(2000)).unwrap();
+
+            let ephemeral_scheduled_fn = ScheduledFn::new("foo".into(), "bar".into());
+            let persisted_scheduled_fn = ScheduledFn::new("1".into(), "2".into());
+            let persisted_schedule = Schedule::Persisted("* * * * * * *".into());
+
+            schedule_fn(
+                txn,
+                &alice_pubkey,
+                persisted_scheduled_fn.clone(),
+                Some(persisted_schedule.clone()),
+                now,
+            )
+            .unwrap();
+            schedule_fn(
+                txn,
+                &alice_pubkey,
+                ephemeral_scheduled_fn.clone(),
+                None,
+                now,
+            )
+            .unwrap();
+
+            assert!(fn_is_scheduled(txn, persisted_scheduled_fn.clone(), &alice_pubkey).unwrap());
+            assert!(fn_is_scheduled(txn, ephemeral_scheduled_fn.clone(), &alice_pubkey).unwrap());
+
+            // Deleting live ephemeral scheduled fns from now should delete.
+            delete_live_ephemeral_scheduled_fns(txn, now, &alice_pubkey).unwrap();
+            assert!(!fn_is_scheduled(txn, ephemeral_scheduled_fn.clone(), &alice_pubkey,).unwrap());
+            assert!(fn_is_scheduled(txn, persisted_scheduled_fn.clone(), &alice_pubkey,).unwrap());
+
+            schedule_fn(
+                txn,
+                &alice_pubkey,
+                ephemeral_scheduled_fn.clone(),
+                None,
+                now,
+            )
+            .unwrap();
+            assert!(fn_is_scheduled(txn, ephemeral_scheduled_fn.clone(), &alice_pubkey,).unwrap());
+            assert!(fn_is_scheduled(txn, persisted_scheduled_fn.clone(), &alice_pubkey,).unwrap());
+
+            // Deleting live ephemeral fns from a past time should do nothing.
+            delete_live_ephemeral_scheduled_fns(txn, the_past, &alice_pubkey).unwrap();
+            assert!(fn_is_scheduled(txn, ephemeral_scheduled_fn.clone(), &alice_pubkey,).unwrap());
+            assert!(fn_is_scheduled(txn, persisted_scheduled_fn.clone(), &alice_pubkey,).unwrap());
+
+            // Deleting live ephemeral fns from the future should delete.
+            delete_live_ephemeral_scheduled_fns(txn, the_future, &alice_pubkey).unwrap();
+            assert!(!fn_is_scheduled(txn, ephemeral_scheduled_fn.clone(), &alice_pubkey,).unwrap());
+            assert!(fn_is_scheduled(txn, persisted_scheduled_fn.clone(), &alice_pubkey,).unwrap());
+
+            // Deleting all ephemeral fns should delete.
+            schedule_fn(
+                txn,
+                &alice_pubkey,
+                ephemeral_scheduled_fn.clone(),
+                None,
+                now,
+            )
+            .unwrap();
+            assert!(fn_is_scheduled(txn, ephemeral_scheduled_fn.clone(), &alice_pubkey,).unwrap());
+            delete_all_ephemeral_scheduled_fns(txn).unwrap();
+            assert!(!fn_is_scheduled(txn, ephemeral_scheduled_fn.clone(), &alice_pubkey,).unwrap());
+            assert!(fn_is_scheduled(txn, persisted_scheduled_fn.clone(), &alice_pubkey,).unwrap());
+
+            let ephemeral_future_schedule =
+                Schedule::Ephemeral(std::time::Duration::from_millis(1001));
+            schedule_fn(
+                txn,
+                &alice_pubkey,
+                ephemeral_scheduled_fn.clone(),
+                Some(ephemeral_future_schedule.clone()),
+                now,
+            )
+            .unwrap();
+            assert_eq!(
+                vec![(
+                    persisted_scheduled_fn.clone(),
+                    Some(persisted_schedule.clone()),
+                    false,
+                )],
+                live_scheduled_fns(txn, the_future, &alice_pubkey,).unwrap(),
+            );
+            assert_eq!(
+                vec![
+                    (persisted_scheduled_fn, Some(persisted_schedule), false),
+                    (
+                        ephemeral_scheduled_fn,
+                        Some(ephemeral_future_schedule),
+                        true
+                    ),
+                ],
+                live_scheduled_fns(txn, the_distant_future, &alice_pubkey,).unwrap(),
+            );
+
+            Result::<(), DatabaseError>::Ok(())
+        })
+        .await
+        .unwrap();
+    Ok(())
+}
+
+#[tokio::test(flavor = "multi_thread")]
+#[cfg(feature = "test_utils")]
+async fn schedule_test_wasm() -> anyhow::Result<()> {
+    holochain_trace::test_run();
+    let RibosomeTestFixture {
+        conductor,
+        alice,
+        alice_pubkey,
+        alice_host_fn_caller,
+        bob,
+        bob_pubkey,
+        bob_host_fn_caller,
+        ..
+    } = RibosomeTestFixture::new(TestWasm::Schedule).await;
+
+    // We don't want the scheduler running and messing with our calculations.
+    conductor
+        .start_scheduler(std::time::Duration::from_millis(1_000_000_000))
+        .await?;
+
+    // At first nothing has happened because init won't run until some zome
+    // call runs.
+    let query_tick: Vec<Record> = conductor.call(&alice, "query_tick_init", ()).await;
+    assert!(query_tick.is_empty());
+
+    // Wait to make sure we've init, but it should have happened for sure.
+    while {
+        !alice_host_fn_caller
+            .authored_db
+            .write_async({
+                let alice_pubkey = alice_pubkey.clone();
+                move |txn| {
+                    let persisted_scheduled_fn = ScheduledFn::new(
+                        TestWasm::Schedule.into(),
+                        "cron_scheduled_fn_init".into(),
+                    );
+
+                    Result::<bool, DatabaseError>::Ok(
+                        fn_is_scheduled(txn, persisted_scheduled_fn.clone(), &alice_pubkey)
+                            .unwrap(),
+                    )
+                }
+            })
+            .await
+            .unwrap()
+    } {
+        tokio::time::sleep(std::time::Duration::from_millis(1)).await;
+    }
+
+    // Round up to the next second so we don't trigger two tocks in quick
+    // succession.
+    let mut now =
+        Timestamp::from_micros((Timestamp::now().as_micros() / 1_000_000 + 1) * 1_000_000 + 1);
+
+    // The ephemeral function will dispatch each millisecond.
+    // The tock will dispatch once and wait a second.
+    let mut i: usize = 0;
+    while i < 10 {
+        conductor.dispatch_scheduled_fns(now).await;
+        now = (now + std::time::Duration::from_millis(2))?;
+        i += 1;
+    }
+    loop {
+        let query_tick_init: Vec<Record> = conductor.call(&alice, "query_tick_init", ()).await;
+        let query_tock_init: Vec<Record> = conductor.call(&alice, "query_tock_init", ()).await;
+        if query_tick_init.len() == 5 && query_tock_init.len() == 1 {
+            break;
+        }
+    }
+
+    // after a second the tock will run again.
+    now = (now + std::time::Duration::from_millis(1000))?;
+    conductor.dispatch_scheduled_fns(now).await;
+    loop {
+        let query_tick_init: Vec<Record> = conductor.call(&alice, "query_tick_init", ()).await;
+        let query_tock_init: Vec<Record> = conductor.call(&alice, "query_tock_init", ()).await;
+        if query_tick_init.len() == 5 && query_tock_init.len() == 2 {
+            break;
+        }
+    }
+
+    // alice can schedule things outside of init.
+    let query_tock: Vec<Record> = conductor.call(&alice, "query_tock", ()).await;
+    assert!(query_tock.is_empty());
+
+    let _schedule: () = conductor.call(&alice, "schedule", ()).await;
+
+    // Round up to the next second so we don't trigger two tocks in quick
+    // succession.
+    now = Timestamp::from_micros((Timestamp::now().as_micros() / 1_000_000 + 1) * 1_000_000 + 1);
+
+    let mut i: usize = 0;
+    while i < 10 {
+        conductor.dispatch_scheduled_fns(now).await;
+        now = (now + std::time::Duration::from_millis(2))?;
+        i += 1;
+    }
+    loop {
+        tokio::time::sleep(std::time::Duration::from_millis(1)).await;
+        let query_tick: Vec<Record> = conductor.call(&alice, "query_tick", ()).await;
+        let query_tock: Vec<Record> = conductor.call(&alice, "query_tock", ()).await;
+        if query_tick.len() == 5 && query_tock.len() == 1 {
+            break;
+        }
+    }
+
+    // after a second the tock will run again.
+    now = (now + std::time::Duration::from_millis(1000))?;
+    conductor.dispatch_scheduled_fns(now).await;
+    loop {
+        tokio::time::sleep(std::time::Duration::from_millis(1)).await;
+        let query_tick: Vec<Record> = conductor.call(&alice, "query_tick", ()).await;
+        let query_tock: Vec<Record> = conductor.call(&alice, "query_tock", ()).await;
+        if query_tick.len() == 5 && query_tock.len() == 2 {
+            break;
+        }
+    }
+
+    // Starting the scheduler should flush ephemeral.
+    let _schedule: () = conductor.call(&bob, "schedule", ()).await;
+
+    assert!(bob_host_fn_caller
+        .authored_db
+        .write_async({
+            let bob_pubkey = bob_pubkey.clone();
+            move |txn| {
+                let persisted_scheduled_fn =
+                    ScheduledFn::new(TestWasm::Schedule.into(), "scheduled_fn".into());
+
+                Result::<bool, DatabaseError>::Ok(
+                    fn_is_scheduled(txn, persisted_scheduled_fn.clone(), &bob_pubkey).unwrap(),
+                )
+            }
+        })
+        .await
+        .unwrap());
+
+    conductor
+        .start_scheduler(std::time::Duration::from_millis(1_000_000_000))
+        .await?;
+
+    assert!(!bob_host_fn_caller
+        .authored_db
+        .write_async({
+            let bob_pubkey = bob_pubkey.clone();
+            move |txn| {
+                let persisted_scheduled_fn =
+                    ScheduledFn::new(TestWasm::Schedule.into(), "scheduled_fn".into());
+
+                Result::<bool, DatabaseError>::Ok(
+                    fn_is_scheduled(txn, persisted_scheduled_fn.clone(), &bob_pubkey).unwrap(),
+                )
+            }
+        })
+        .await
+        .unwrap());
+
+    Ok(())
 }
